Create RViz visualizer script
Create a ROS 2 script that sends the polygon poses to RViz as
markers, as a way to visually confirm the function is working
correctly.

Attempted to create a script that uses the Nav2 framework to
have a robot navigate to the polygon poses, but was not able to
get it to work.

Updated test_pose_generators.py to test for polygons of various
number of sides, and also confirm the yaw change between
poses is correct.